+++ services/product-service/src/main/java/com/vmarket/product/service/ProductDerivedFields.java
@@ -1,6 +1,5 @@
 package com.vmarket.product.service;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.stereotype.Component;
@@ -13,9 +12,9 @@ public class ProductDerivedFields {
 	public void refresh(Product product) {
 		List<ProductVariant> variants = product.getVariants() == null ? List.of() : product.getVariants();
 		product.setMinPrice(variants.stream().map(ProductVariant::getPrice)
-				.min(BigDecimal::compareTo).orElse(null));
+				.min(Long::compareTo).orElse(null));
 		product.setMaxPrice(variants.stream().map(ProductVariant::getPrice)
-				.max(BigDecimal::compareTo).orElse(null));
+				.max(Long::compareTo).orElse(null));
 		product.setAvailableStock(variants.stream()
 				.mapToLong(variant -> Math.max(0, variant.getStock() - variant.getReservedStock())).sum());
 	}
